Renames icon files
